Add CMake configuration
Unfortunately this time has come ;C
All the libraries I use are configured with CMake. Every time I want
to use some libraries, I need to prepare Premake5 scripts myself and,
even worse, maintain them. Not only for Windows but also for Linux.
It is just too much for one person.

CMake also enables me to use CLion and other IDEs on non-Windows
platforms. I really love Premake for the clarity of projects that
it generates but its adoption is not enough for me to bear.